+++ src/main/java/acme/features/employer/duty/EmployerDutyController.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import acme.entities.dutys.Duty;
+import acme.entities.duties.Duty;
 import acme.entities.roles.Employer;
 import acme.framework.components.BasicCommand;
 import acme.framework.controllers.AbstractController;
@@ -19,14 +19,18 @@ public class EmployerDutyController extends AbstractController<Employer, Duty> {
 	// Internal state ---------------------------------------------------------
 
 	@Autowired
-	private EmployerDutyCreateService createService;
+	private EmployerDutyListService	listService;
+
+	@Autowired
+	private EmployerDutyShowService		showService;
 
 
 	// Constructors -----------------------------------------------------------
 
 	@PostConstruct
 	private void initialise() {
-		super.addBasicCommand(BasicCommand.CREATE, this.createService);
+		super.addBasicCommand(BasicCommand.LIST, this.listService);
+		super.addBasicCommand(BasicCommand.SHOW, this.showService);
 
 	}
 }
