--- src/main/java/acme/features/employer/duty/EmployerDutyCreateService.java
@@ -1,71 +1,0 @@
-
-package acme.features.employer.duty;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import acme.entities.dutys.Duty;
-import acme.entities.roles.Employer;
-import acme.framework.components.Errors;
-import acme.framework.components.Model;
-import acme.framework.components.Request;
-import acme.framework.services.AbstractCreateService;
-
-@Service
-public class EmployerDutyCreateService implements AbstractCreateService<Employer, Duty> {
-
-	// Internal state ---------------------------------------------------------
-
-	@Autowired
-	EmployerDutyRepository repository;
-
-
-	// AbstractListService<Employer, Duty> interface --------------------
-
-	@Override
-	public boolean authorise(final Request<Duty> request) {
-		assert request != null;
-
-		return true;
-	}
-
-	@Override
-	public void bind(final Request<Duty> request, final Duty entity, final Errors errors) {
-		assert request != null;
-		assert entity != null;
-		assert errors != null;
-
-		request.bind(entity, errors, "moment");
-	}
-
-	@Override
-	public void unbind(final Request<Duty> request, final Duty entity, final Model model) {
-		assert request != null;
-		assert entity != null;
-		assert model != null;
-
-		request.unbind(entity, model, "title", "description", "percentage");
-
-	}
-
-	@Override
-	public Duty instantiate(final Request<Duty> request) {
-		Duty result;
-
-		result = new Duty();
-		return result;
-	}
-
-	@Override
-	public void validate(final Request<Duty> request, final Duty entity, final Errors errors) {
-		assert request != null;
-		assert entity != null;
-		assert errors != null;
-	}
-	@Override
-	public void create(final Request<Duty> request, final Duty entity) {
-
-		this.repository.save(entity);
-	}
-
-}
