+++ src/main/java/acme/features/employer/duty/EmployerDutyRepository.java
@@ -6,7 +6,7 @@ import java.util.Collection;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import acme.entities.dutys.Duty;
+import acme.entities.duties.Duty;
 import acme.framework.repositories.AbstractRepository;
 
 @Repository
