+++ src/main/java/acme/features/employer/duty/EmployerDutyShowService.java
@@ -1,0 +1,51 @@
+
+package acme.features.employer.duty;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import acme.entities.duties.Duty;
+import acme.entities.roles.Employer;
+import acme.framework.components.Model;
+import acme.framework.components.Request;
+import acme.framework.services.AbstractShowService;
+
+@Service
+public class EmployerDutyShowService implements AbstractShowService<Employer, Duty> {
+	// Internal state ---------------------------------------------------------
+
+	@Autowired
+	EmployerDutyRepository repository;
+
+
+	@Override
+	public boolean authorise(final Request<Duty> request) {
+		assert request != null;
+
+		return true;
+	}
+
+	@Override
+	public void unbind(final Request<Duty> request, final Duty entity, final Model model) {
+		assert request != null;
+		assert entity != null;
+		assert model != null;
+
+		request.unbind(entity, model, "title", "description", "percentage");
+
+	}
+
+	@Override
+	public Duty findOne(final Request<Duty> request) {
+		assert request != null;
+
+		Duty result;
+		int id;
+
+		id = request.getModel().getInteger("id");
+
+		result = this.repository.findOnebyId(id);
+
+		return result;
+	}
+}
