+++ src/main/webapp/META-INF/views/employer/duty/i18n_en.messages
@@ -14,8 +14,13 @@
 employer.duty.form.label.title = Title
 employer.duty.form.label.description = Description
 employer.duty.form.label.percentage = Percentage
-employer.duty.form.button.create = Create
 employer.duty.form.button.return= Return
+
+
+#DUTY LIST
+employer.duty.list.label.title= Title
+employer.duty.list.label.description= Description
+employer.duty.list.label.percentage= Percentage
