+++ src/main/webapp/META-INF/views/employer/duty/i18n_es.messages
@@ -13,6 +13,12 @@
 
 employer.duty.form.label.title = Título
 employer.duty.form.label.description = Descripción
-employer.duty.form.label.percentage = Porcentage
+employer.duty.form.label.percentage = Porcentaje
 employer.duty.form.button.create = Crear
 employer.duty.form.button.return= Volver
+
+#DUTY LIST
+employer.duty.list.label.title= Título
+employer.duty.list.label.description= Descripción
+employer.duty.list.label.percentage= Porcentaje
+
