+++ src/main/webapp/META-INF/views/employer/duty/list.jsp
@@ -6,7 +6,7 @@
 - In keeping with the traditional purpose of furthering education and research, it is
 - the policy of the copyright owner to permit non-commercial use and redistribution of
 - this software. It has been tested carefully, but it is not guaranteed for any particular
-- purposes.  The copyright owner does not duty any warranties or representations, nor do
+- purposes.  The copyright owner does not offer any warranties or representations, nor do
 - they accept any liabilities with respect to them.
 --%>
 
@@ -15,9 +15,10 @@
 <%@taglib prefix="jstl" uri="http://java.sun.com/jsp/jstl/core"%>
 <%@taglib prefix="acme" tagdir="/WEB-INF/tags"%>
 
-<acme:form>
-	<acme:form-textbox code="employer.duty.form.label.title" path="title"/>
-	<acme:form-textarea code="employer.duty.form.label.description" path="description"/>
-	<acme:form-integer code="employer.duty.form.label.percentage" path="percentage"/>
-	<acme:form-return code="employer.duty.form.button.return"/>
-</acme:form>
+
+<acme:list>
+	<acme:list-column code="employer.duty.list.label.title" path="title" width="10%"/>
+	<acme:list-column code="employer.duty.list.label.description" path="description" width="10%"/>
+	<acme:list-column code="employer.duty.list.label.percentage" path="percentage" width="10%"/>
+
+</acme:list>
