+++ src/main/webapp/META-INF/views/master/i18n_en.messages
@@ -84,7 +84,7 @@ master.menu.worker.application.list= List applications
 # master.menu.employer ########################################################
 
 master.menu.employer = Employer
-master.menu.employer.duty.create = Create duty
+master.menu.employer.duty.list = List duties
 master.menu.employer.job.list= List jobs
 master.menu.employer.application.list= List applications
