+++ src/main/webapp/META-INF/views/master/menu.jsp
@@ -86,7 +86,7 @@
 	  <%-- COSAS DE EMPLOYER --%>
 
 		<acme:menu-option code="master.menu.employer" access="hasRole('Employer')">
-			<acme:menu-suboption code="master.menu.employer.duty.create" action="/employer/duty/create"/>
+			<acme:menu-suboption code="master.menu.employer.duty.list" action="/employer/duty/list"/>
 			<acme:menu-suboption code="master.menu.employer.job.list" action="/employer/job/list-mine"/>
 			<acme:menu-suboption code="master.menu.employer.application.list" action="/employer/application/list-mine"/>
 		</acme:menu-option>
